refactor: fetch json document from url
